checkAuth pulling out of authentication route file to become middleware - as intended
