+++ config/middleware/checkAuth.js
@@ -1,0 +1,18 @@
+const passport = require('../../config/passport-config');
+
+// middleware
+const checkAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/login');
+  }
+  
+  const checkNotAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return res.redirect('/');
+    }
+    next();
+  }
+  
+  module.exports = { checkAuthenticated, checkNotAuthenticated }
